# Speed-up HTML escaping a little bit